# packages/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Hookwing Webhook API
  version: "0.0.1"
  description: |
    Hookwing is a webhook delivery platform that reliably receives, routes, and delivers webhooks
    with automatic retries, signature verification, and tier-based feature gating.
  contact:
    name: Hookwing Engineering
    url: https://hookwing.com
    email: support@hookwing.com
servers:
  - url: https://api.hookwing.com
    description: Production
  - url: https://api-dev.hookwing.com
    description: Development
security:
  - BearerAuth: []
...
paths:
...
  /v1/events:
    get:
      operationId: listEvents
      tags: [Events]
      summary: List events (filtered, cursor-paginated)
      description: |
        Returns events for your workspace with cursor-based pagination and optional filters.
        Events older than your tier's retention period are automatically excluded.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          description: Event ID cursor for pagination (events before this ID)
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: event_type
          in: query
          schema:
            type: string
        - name: since
          in: query
          description: ISO 8601 timestamp — events received after this time
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          description: ISO 8601 timestamp — events received before this time
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Events list with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/events/{id}:
    get:
      operationId: getEvent
      tags: [Events]
      summary: Get event with delivery details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event with deliveries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/events/{id}/replay:
    post:
      operationId: replayEvent
      tags: [Events]
      summary: Replay a single event
      description: Re-delivers the event to all active endpoints in the workspace.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplayResponse"
        "400":
          description: No active endpoints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/events/replay:
    post:
      operationId: bulkReplayEvents
      tags: [Events]
      summary: Bulk replay events (up to 100)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkReplayInput"
            example:
              eventIds:
                - evt_01HXYZ123
                - evt_01HXYZ456
      responses:
        "200":
          description: Events replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkReplayResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ==========================================================================
  # Deliveries
  # ==========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key from signup or key creation. Format - `hk_live_<random>`
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Unauthorized
            message: Missing Authorization header
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Not found
...
    ValidationError:
      description: Invalid request body
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid input
              details:
                type: object
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
...
    Event:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        eventType:
          type: string
        payload:
          type: object
          nullable: true
        headers:
          type: object
          nullable: true
        sourceIp:
          type: string
          nullable: true
        receivedAt:
          type: integer
        processedAt:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, processing, completed, failed]
    EventDetail:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          properties:
            deliveries:
              type: array
              items:
                $ref: "#/components/schemas/Delivery"
    EventListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        pagination:
          $ref: "#/components/schemas/CursorPagination"
    CursorPagination:
      type: object
      properties:
        limit:
          type: integer
        cursor:
          type: string
          nullable: true
          description: Pass as cursor param to get next page
        hasMore:
          type: boolean
    ReplayResponse:
      type: object
      properties:
        replayed:
          type: boolean
        eventId:
          type: string
        deliveryIds:
          type: array
          items:
            type: string
        endpointCount:
          type: integer
    BulkReplayInput:
      type: object
      required: [eventIds]
      properties:
        eventIds:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 100
    BulkReplayResponse:
      type: object
      properties:
        replayed:
          type: boolean
        count:
          type: integer
        deliveryIds:
          type: array
          items:
            type: string
        endpointCount:
          type: integer

    # -- Deliveries --
    Delivery:
      type: object
      properties:
        id:
          type: string
        endpointId:
          type: string
        eventId:
          type: string
        attemptNumber:
          type: integer
        status:
          type: string
          enum: [pending, success, failed, retrying]
        responseStatusCode:
          type: integer
          nullable: true
        responseBody:
          type: string
          nullable: true
          description: First 1KB of response body
        responseHeaders:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
        durationMs:
          type: integer
          nullable: true
        nextRetryAt:
          type: integer
          nullable: true
        deliveredAt:
          type: integer
          nullable: true
        createdAt:
          type: integer
...